# Frame: Add File, Func, Line methods