--- a/format_test.go
+++ b/format_test.go
@@ -533,3 +533,7 @@ func testGenericRecursive(t *testing.T, beforeErr error, beforeWant []string, li
 		}
 	}
 }
+
+func regexpMatchString(pattern, s string) (bool, error) {
+	return regexp.MatchString(pattern, s)
+}