Implementation of proposal in #99.

Benchmarks, for good measure:

Old:

```
$ go test -bench .
BenchmarkErrors/pkg/errors-stack-10-8            1000000              1820 ns/op             320 B/op          3 allocs/op
BenchmarkErrors/errors-stack-10-8               20000000                72.2 ns/op            16 B/op          1 allocs/op
BenchmarkErrors/pkg/errors-stack-100-8            500000              3249 ns/op             320 B/op          3 allocs/op
BenchmarkErrors/errors-stack-100-8               3000000               513 ns/op              16 B/op          1 allocs/op
BenchmarkErrors/pkg/errors-stack-1000-8           200000             11130 ns/op             320 B/op          3 allocs/op
BenchmarkErrors/errors-stack-1000-8               300000              4877 ns/op              16 B/op          1 allocs/op
PASS
ok      github.com/pkg/errors   11.294s
```

New:

```
$ go test -bench .
BenchmarkErrors/pkg/errors-stack-10-8            1000000              1825 ns/op             320 B/op          3 allocs/op
BenchmarkErrors/errors-stack-10-8               20000000                73.0 ns/op            16 B/op          1 allocs/op
BenchmarkErrors/pkg/errors-stack-100-8            500000              3229 ns/op             320 B/op          3 allocs/op
BenchmarkErrors/errors-stack-100-8               3000000               509 ns/op              16 B/op          1 allocs/op
BenchmarkErrors/pkg/errors-stack-1000-8           200000             11092 ns/op             320 B/op          3 allocs/op
BenchmarkErrors/errors-stack-1000-8               300000              4867 ns/op              16 B/op          1 allocs/op
PASS
ok      github.com/pkg/errors   11.288s
```

/cc @davecheney @ChrisHines