@davecheney 2017-01-18
Sorry I haven't had a chance to look at this yet; I hope to have some free
time on the weekend.

On Thu, Jan 19, 2017 at 4:33 AM, Jud White <notifications@github.com> wrote:

> @davecheney <https://github.com/davecheney> Would you mind providing your
> feedback?
>
> —
> You are receiving this because you were mentioned.
> Reply to this email directly, view it on GitHub
> <https://github.com/pkg/errors/pull/100#issuecomment-273543897>, or mute
> the thread
> <https://github.com/notifications/unsubscribe-auth/AAAcA1LHY4FVGyCMiNzthN6Y9G1u1CP3ks5rTkzggaJpZM4Lj4cu>
> .
>