@judwhite 2017-02-19
@davecheney Friendly ping. If you want I can work on #38, although I'd still like to get specific values out of pkg/errors so users don't have to re-parse the text if they want these values on their own.