@davecheney 2017-02-25
Sorry for the extremely long delay.

I'm not sure I want to add this because it is going beyond the remit of this package. If you have the `Frame` value then you have the pc within a function and can use `runtime.FuncForPc` and friends.

If I had a choice I would not have exported `Frame`, but it was necessary to make the value implement `fmt.Formatter` _and_ still be convertible to a `uintptr`.

I'm wary about adding these methods on top of Frame because they are redundant, see above, but more seriously I'm concerned they'll be a precedent for adding more methods. This was the reason for going down the fmt.Formatter route, so the Error and Frame types would not grow a forrest of helper methods to introspect them.